# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Accounting Platform API
  version: 1.0.0
  description: |
    Multi-tenant accounting platform. All business resources are tenant-scoped and most are
    legal-entity scoped; scope is taken from the authenticated session, never from a request
    parameter. Money is carried as decimal strings, never JSON numbers (ADR-0006).
  x-supersedes:
    - project plan/blueprints/api_route_catalog.yaml
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: https://sandbox.api.example.com/v1
    description: Sandbox (doc 15 developer platform)
security:
  - sessionBearer: []
  - oauth2: []
...
paths:
...
  /tasks:            {get:  {operationId: listTasks,    tags: [Workflow], x-permission: task.view,   x-idempotency: n-a,      x-phase: 6, x-added-by: F-106, description: "Unified inbox across 10 source types (doc 14).", responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
  /tasks/{id}/complete: {post: {operationId: completeTask, tags: [Workflow], x-permission: task.complete, x-idempotency: required, x-phase: 6, parameters: [{$ref: '#/components/parameters/Id'}], responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
...
  /close-runs:
    get:  {operationId: listCloseRuns,  tags: [Close], x-permission: close.view,   x-idempotency: n-a,         x-phase: 6, responses: {'200': {$ref: '#/components/responses/Ok'},      default: {$ref: '#/components/responses/Error'}}}
    post: {operationId: createCloseRun, tags: [Close], x-permission: close.manage, x-idempotency: recommended, x-phase: 6, responses: {'201': {$ref: '#/components/responses/Created'}, default: {$ref: '#/components/responses/Error'}}}
  /close-runs/{id}/tasks:    {get:  {operationId: listCloseTasks,  tags: [Close], x-permission: close.view,   x-idempotency: n-a,      x-phase: 6, parameters: [{$ref: '#/components/parameters/Id'}], responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
  /close-runs/{id}/finalize: {post: {operationId: finalizeCloseRun, tags: [Close], x-permission: close.finalize, x-idempotency: required, x-phase: 6, description: "Blocked while required tasks are incomplete unless an override workflow exists (doc 12).", parameters: [{$ref: '#/components/parameters/Id'}], responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
  /close-runs/{id}/reopen:   {post: {operationId: reopenCloseRun,   tags: [Close], x-permission: close.reopen,   x-idempotency: required, x-phase: 6, x-high-risk: true, x-added-by: F-1002, description: "doc 21 Phase 6 requires a \"reopen workflow with privileged approval/audit\" and its exit criterion is that \"reopening a period is privileged, approved and audited\"; Gate F tests the same thing. close_runs has carried reopened_at, reopened_by and reopen_reason since 0016, and F-038 deliberately made a reopen a new close VERSION rather than an edit - and none of it was reachable, so a completed close was final regardless of what was found afterwards. Answers 201 because the reopen creates the next version rather than editing the closed one: the completed run moves to REOPENED - which is also what frees close_runs_completed_uq for the version that follows - and version+1 opens beside it, so the previous version's task history survives instead of being overwritten by the second attempt at the same close. Its own permission code, not close.finalize and not accounting_period.reopen: finalize and reopen are the two halves Gate F insists be separable, and folding this into the period permission would mean everyone who may reopen a calendar period may also supersede a completed close (F-921). Flagged high-risk under doc 02's \"reopen periods\" category, so it demands re-authentication. Refused unless the reopen has been approved through the approval engine by somebody other than whoever finalized the close, and the reason is recorded - close_runs_reopen_has_reason makes the reason mandatory in the database, and the approval is what makes this a workflow rather than a privilege.", parameters: [{$ref: '#/components/parameters/Id'}], responses: {'201': {$ref: '#/components/responses/Created'}, default: {$ref: '#/components/responses/Error'}}}}
  /close-tasks/{id}/complete: {post: {operationId: completeCloseTask, tags: [Close], x-permission: close.manage, x-idempotency: required, x-phase: 6, parameters: [{$ref: '#/components/parameters/Id'}], responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
  /close-checklists/generate: {post: {operationId: generateCloseChecklist, tags: [Close], x-permission: close.manage, x-idempotency: required, x-phase: 6, x-added-by: F-106, responses: {'201': {$ref: '#/components/responses/Created'}, default: {$ref: '#/components/responses/Error'}}}}
  /account-reconciliations:
    get:  {operationId: listAccountReconciliations,  tags: [Close], x-permission: reconciliation.view,   x-idempotency: n-a,         x-phase: 6, x-added-by: F-106, description: "What Gate D tests: GL balance, subledger balance, reconciling items, difference.", responses: {'200': {$ref: '#/components/responses/Ok'},      default: {$ref: '#/components/responses/Error'}}}
    post: {operationId: createAccountReconciliation, tags: [Close], x-permission: reconciliation.manage, x-idempotency: recommended, x-phase: 6, x-added-by: F-106, responses: {'201': {$ref: '#/components/responses/Created'}, default: {$ref: '#/components/responses/Error'}}}
  /account-reconciliations/{id}/certify: {post: {operationId: certifyAccountReconciliation, tags: [Close], x-permission: reconciliation.certify, x-idempotency: required, x-phase: 6, x-added-by: F-106, parameters: [{$ref: '#/components/parameters/Id'}], responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
  /account-reconciliations/{id}/reconciling-items:
    get:  {operationId: listReconcilingItems,  tags: [Close], x-permission: reconciliation.view,   x-idempotency: n-a,      x-phase: 6, x-added-by: F-1001, description: "Gate D: \"Any difference is either zero or represented by an explicit, explainable reconciliation item.\" reconciling_items has held that explanation since 0016 and no route could read or write one, which left the gate provable only in its trivial direction: account_reconciliations_explained compares the difference against the sum of the items, so with no items a zero difference certified while stating nothing, and a non-zero difference could never be certified at all. This is the read half of the artifact the gate asserts on - type, amount, expected clear date, evidence and whether it has cleared.", parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}], responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}
    post: {operationId: createReconcilingItem, tags: [Close], x-permission: reconciliation.manage, x-idempotency: required, x-phase: 6, x-added-by: F-1001, description: "The write half, and what makes a non-zero difference certifiable: the deferred constraint trigger on account_reconciliations requires the items to sum exactly to the difference, so the difference between a subledger and its control account is explained item by item or the certification is refused. Same permission as createAccountReconciliation on purpose - preparing the reconciliation and stating why it differs are one act by one preparer, and certification is the separate control (reconciliation.certify, with recon_sod requiring a different person). Idempotency-Key required rather than recommended: a retried create adds a second item with the same amount, which over-explains the difference and is then caught at certification - correct, but it costs a preparer an afternoon hunting an item they never knowingly created.", parameters: [{$ref: '#/components/parameters/Id'}], responses: {'201': {$ref: '#/components/responses/Created'}, default: {$ref: '#/components/responses/Error'}}}
  /account-reconciliations/{id}/reopen:   {post: {operationId: reopenAccountReconciliation,   tags: [Close], x-permission: reconciliation.reopen,   x-idempotency: required, x-phase: 6, x-high-risk: true, x-added-by: F-1001, description: "doc 12: \"Approved reconciliation is immutable unless formally reopened with reason.\" recon_cert_status has carried REOPENED and account_reconciliations a reopened_reason from 0016, and no route could reach either - so a certified reconciliation whose items turned out to be wrong had no lawful correction path at all, which is precisely the condition under which somebody reaches for the DELETE privilege on reconciling_items instead. Added with the reconciling-item routes rather than after them: shipping the write path for the explanation without a way to withdraw a certification designs people into that hole. The banking twin, reopenBankReconciliation, has existed since Phase 4 and is flagged the same way; this is the same act on the subledger-to-control reconciliation Gate D reads.", parameters: [{$ref: '#/components/parameters/Id'}], responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
  /reconciling-items/{id}/resolve: {post: {operationId: resolveReconcilingItem, tags: [Close], x-permission: reconciliation.manage, x-idempotency: required, x-phase: 6, x-added-by: F-1001, description: "doc 12 tracks a reconciling item to its clearance: expected_clear_date and resolved_at exist so a timing difference still open past the date it was expected to clear becomes the exception Gate D wants \"visible before close\", rather than a line that sits in the list forever and is read as normal. Nothing could set resolved_at, so every item was permanently outstanding and the exception list was everything or nothing.", parameters: [{$ref: '#/components/parameters/Id'}], responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
  /financial-statement-snapshots:
    get:  {operationId: listFinancialSnapshots,  tags: [Close], x-permission: report.view,     x-idempotency: n-a,      x-phase: 6, x-added-by: F-106, responses: {'200': {$ref: '#/components/responses/Ok'},      default: {$ref: '#/components/responses/Error'}}}
    post: {operationId: createFinancialSnapshot, tags: [Close], x-permission: report.snapshot, x-idempotency: required, x-phase: 6, x-added-by: F-106, description: "Immutable once created; future activity must not rewrite it (Gate E).", responses: {'201': {$ref: '#/components/responses/Created'}, default: {$ref: '#/components/responses/Error'}}}
  /financial-statement-snapshots/{id}/approve: {post: {operationId: approveFinancialSnapshot, tags: [Close], x-permission: financial_statement.approve, x-idempotency: required, x-phase: 6, x-added-by: F-1003, description: "doc 12: \"When statements are approved/period closed, save immutable metadata: report definition version, parameters, ledger cutoff/checkpoint, generated file/hash, actor, approval and date\", and contracts/events.yaml has defined financial_statements.approved from the start with nothing able to emit it. financial_snapshots.approved_by and approved_at have never been writable by anyone: the row is written once by createFinancialSnapshot, and the 0016 immutability trigger refuses every UPDATE unconditionally while Gate C revokes the privilege as well - so statement approval was structurally unreachable rather than merely unbuilt. The approval must therefore be recorded as a separate append-only fact; neither of those two controls may be relaxed to make this route work, because Gate E's \"snapshot is unchanged by future transactions\" and Gate C's revocation are deliberately independent. Its own resource and permission because the person who issues a statement is not the person who approves it, and sharing report.snapshot would have made that separation ungrantable (F-921); the approver being different from the snapshot's creator is a rule the recording must enforce NULL-safely, since a comparison against a null creator is null and satisfies a CHECK.", parameters: [{$ref: '#/components/parameters/Id'}], responses: {'201': {$ref: '#/components/responses/Created'}, default: {$ref: '#/components/responses/Error'}}}}
...
  /period-end-entries/{id}: {get: {operationId: getPeriodEndEntry, tags: [Close], x-permission: period_end_entry.view, x-idempotency: n-a, x-phase: 6, x-added-by: F-1004, description: "The entry, its recognition schedule and the journals it has produced so far. A prepaid spread over twelve months is twelve postings against one balance and a list row cannot show whether the eleventh happened; this is the drill behind the checklist task and the evidence that the balance remaining is the balance the schedule says should remain.", parameters: [{$ref: '#/components/parameters/Id'}], responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
...
components:
  securitySchemes:
    sessionBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        First-party session token. Carries a session ID checked against the `sessions` table on
        refresh, so revocation is immediate (ADR-0005). Tenant is bound to the session, not the
        credential.
    oauth2:
      type: oauth2
      description: Third-party applications (doc 15). Scopes map 1:1 to x-permission values.
      flows:
        authorizationCode:
          authorizationUrl: https://api.example.com/oauth/authorize
          tokenUrl: https://api.example.com/oauth/token
          scopes: {}
  parameters:
    Id:         {name: id,     in: path,  required: true,  schema: {type: string, format: uuid}}
...
  responses:
    Ok:        {description: Success,  content: {application/json: {schema: {type: object}}}}
    Created:   {description: Created,  content: {application/json: {schema: {type: object}}}}
...
    Error:
      description: |
        Structured error with a stable code and correlation ID. Notable codes:
        DUPLICATE_SUSPECTED (409, vendor bill override), IN_PROGRESS (409, idempotency),
        PERIOD_CLOSED (422), APPROVAL_REQUIRED (403), UNBALANCED_JOURNAL (422),
        OVER_ALLOCATION (422), INSUFFICIENT_STOCK (422).
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}